[#865] Use HSV for color diff computation in template matching
Unit tests updated for with new image format + new test for this bug
Should be executed connected to an Android device/emulator and executed with
./gradlew :core:smart:detection:connectedFDroidDebugAndroidTest
