+++ core/smart/detection/src/androidTest/java/com/buzbuz/smartautoclicker/core/detection/data/TestImages.kt
@@ -58,6 +58,12 @@ internal sealed class TestImage(@RawRes val fileRes: Int, val size: Point) {
                 ),
             )
         )
+
+        data object TutorialTargetGrayscale : Condition(
+            fileRes = R.raw.condition_1_grayscale,
+            size = Point(198, 192),
+            expectedResults = emptyMap(),
+        )
     }
 }
