+++ core/smart/detection/src/androidTest/java/com/buzbuz/smartautoclicker/core/detection/utils/Constants.kt
@@ -18,3 +18,5 @@ package com.buzbuz.smartautoclicker.core.detection.utils
 
 /** Allow to always returns the best match, even if not up to standards. */
 internal const val TEST_DETECTION_THRESHOLD_ALL = 100
+/** Realistic threshold matching typical user configuration. */
+internal const val TEST_DETECTION_THRESHOLD_STANDARD = 15
