+++ core/smart/detection/src/main/cpp/detector/images/condition_image.hpp
@@ -27,8 +27,6 @@ namespace smartautoclicker {
 
     public:
         void processNewData(std::unique_ptr<cv::Mat> newData, int targetWidth, int targetHeight);
-
-        [[nodiscard]] cv::Scalar getColorMean() const;
     };
 }
