+++ core/smart/detection/src/main/cpp/detector/images/screen_image.hpp
@@ -33,6 +33,7 @@ namespace smartautoclicker {
 
         [[nodiscard]] cv::Mat cropColor(const cv::Rect& roi) const;
         [[nodiscard]] cv::Mat cropGray(const cv::Rect& roi) const;
+        [[nodiscard]] cv::Mat cropHsv(const cv::Rect& roi) const;
     };
 }
